requirements.txt
alembic==1.2.1
aniso8601==8.0.0
attrs==19.2.0
Click==7.0
Flask==1.1.1
Flask-Debug==0.4.3
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
gunicorn==19.9.0
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.2
Mako==1.1.0
MarkupSafe==1.1.1
pyrsistent==0.15.4
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.3
six==1.12.0
SQLAlchemy==1.3.9
Werkzeug==0.16.0
